requirements.txt
fastapi
uvicorn
pydantic
python-multipart
Pillow
opencv-python
numpy
rembg
pytesseract
ultralytics
sqlalchemy
alembic
streamlit
diffusers
transformers
torch
accelerate
